chore: create alerts is updated
